SauceDemo.UI project created. Solution is functional, tests are running
