--- SauceDemo.Tests/Config/appsettings___.json
@@ -1,7 +1,0 @@
-﻿{
-  "TestSettings": {
-    "Browser": "chrome",
-    "BaseUrl": "https://www.saucedemo.com/",
-    "WebDriverWaitSeconds": 10
-  }
-}
